refactor extensions module
