+++ build.sbt
@@ -219,8 +219,8 @@ def multiJvmTestSettings: Seq[Setting[_]] = {
 
   SbtMultiJvm.multiJvmSettings ++
     forkedTests ++
-    // enabling HeaderPlugin in MultiJvm requires two sets of settings.
-    // see https://github.com/sbt/sbt-header/issues/37
+    enabling HeaderPlugin in MultiJvm requires two sets of settings.
+    see https://github.com/sbt/sbt-header/issues/37
     headerSettings(MultiJvm) ++
     automateHeaderSettings(MultiJvm) ++
     Seq(
@@ -365,7 +365,7 @@ val javadslProjects = Seq[ProjectReference](
   `akka-management-javadsl`,
   `akka-discovery-service-locator-javadsl`,
   `cluster-javadsl`,
-  `cluster-extensions-javadsl`,
+  `projection-javadsl`,
   `persistence-javadsl`,
   `persistence-cassandra-javadsl`,
   `persistence-jdbc-javadsl`,
@@ -387,7 +387,7 @@ val scaladslProjects = Seq[ProjectReference](
   `akka-management-scaladsl`,
   `akka-discovery-service-locator-scaladsl`,
   `cluster-scaladsl`,
-  `cluster-extensions-scaladsl`,
+  `projection-scaladsl`,
   `persistence-scaladsl`,
   `persistence-cassandra-scaladsl`,
   `persistence-jdbc-scaladsl`,
@@ -406,9 +406,9 @@ val coreProjects = Seq[ProjectReference](
   `akka-management-core`,
   `akka-discovery-service-locator-core`,
   `cluster-core`,
-  `cluster-extensions`,
   `kafka-client`,
   `kafka-broker`,
+  `projection-core`,
   `persistence-core`,
   `persistence-testkit`,
   `persistence-cassandra-core`,
@@ -767,11 +767,13 @@ lazy val `akka-management-scaladsl` = (project in file("akka-management/scaladsl
 lazy val `cluster-core` = (project in file("cluster/core"))
   .dependsOn(`akka-management-core`)
   .settings(runtimeLibCommon: _*)
+  .settings(multiJvmTestSettings: _*)
   .enablePlugins(RuntimeLibPlugins)
   .settings(
     name := "lagom-cluster-core",
     Dependencies.`cluster-core`
   )
+  .configs(MultiJvm)
 
 lazy val `cluster-javadsl` = (project in file("cluster/javadsl"))
   .dependsOn(`akka-management-javadsl`, `cluster-core`, jackson)
@@ -817,45 +819,48 @@ lazy val `pubsub-scaladsl` = (project in file("pubsub/scaladsl"))
   )
   .configs(MultiJvm)
 
-lazy val `cluster-extensions` = (project in file("cluster/extensions"))
-  .dependsOn(`cluster-core`, logback % Test)
+lazy val `projection-core` = (project in file("projection/core"))
+  .dependsOn(
+    `cluster-core` % "compile;multi-jvm->multi-jvm",
+    logback % Test
+  )
   .settings(runtimeLibCommon: _*)
-//  .settings(mimaSettings(since = version150): _*)
-  .settings(multiJvmTestSettings)
+//  .settings(mimaSettings(since = version160): _*) TODO: enable when 1.6.0 is out
+  .settings(multiJvmTestSettings: _*)
   .settings(Protobuf.settings)
   .enablePlugins(RuntimeLibPlugins)
   .settings(
-    name := "lagom-cluster-extensions",
-    Dependencies.`cluster-extensions`
+    name := "lagom-projection-core",
+    Dependencies.`projection-core`
   )
   .configs(MultiJvm)
 
-lazy val `cluster-extensions-scaladsl` = (project in file("cluster/extensions-scaladsl"))
-  .dependsOn(`cluster-extensions`, `cluster-scaladsl`, logback % Test)
+lazy val `projection-scaladsl` = (project in file("projection/scaladsl"))
+  .dependsOn(`projection-core`, `cluster-scaladsl`, logback % Test)
   .settings(runtimeLibCommon: _*)
-  //  .settings(mimaSettings(since = version150): _*)
+  //  .settings(mimaSettings(since = version160): _*) TODO: enable when 1.6.0 is out
   // https://github.com/lagom/lagom/issues/2045
-  //  .settings(Protobuf.settings) // TODO: promote serialisers for EnsureActive to cluster-extensions
+  //  .settings(Protobuf.settings) // TODO: promote serialisers for EnsureActive to cluster-core
   .enablePlugins(RuntimeLibPlugins)
   .settings(
-    name := "lagom-scaladsl-cluster-extensions",
-    Dependencies.`cluster-extensions-scaladsl`
+    name := "lagom-scaladsl-projection",
+    Dependencies.`projection-scaladsl`
   )
 
-lazy val `cluster-extensions-javadsl` = (project in file("cluster/extensions-javadsl"))
-  .dependsOn(`cluster-extensions`, `cluster-javadsl`, logback % Test)
+lazy val `projection-javadsl` = (project in file("projection/javadsl"))
+  .dependsOn(`projection-core`, `cluster-javadsl`, logback % Test)
   .settings(runtimeLibCommon: _*)
-  //  .settings(mimaSettings(since = version150): _*)
+  //  .settings(mimaSettings(since = version160): _*) TODO: enable when 1.6.0 is out
   // https://github.com/lagom/lagom/issues/2045
-  //  .settings(Protobuf.settings) // TODO: promote serialisers for EnsureActive to cluster-extensions
+  //  .settings(Protobuf.settings) // TODO: promote serialisers for EnsureActive to cluster-core
   .enablePlugins(RuntimeLibPlugins)
   .settings(
-    name := "lagom-javadsl-cluster-extensions",
-    Dependencies.`cluster-extensions-javadsl`
+    name := "lagom-javadsl-projection",
+    Dependencies.`projection-javadsl`
   )
 
 lazy val `persistence-core` = (project in file("persistence/core"))
-  .dependsOn(`cluster-core`, `cluster-extensions` % "compile;test->test", logback % Test)
+  .dependsOn(`cluster-core` % "compile;test->test", logback % Test)
   .settings(runtimeLibCommon: _*)
   .settings(mimaSettings(since = version150): _*)
   .settings(Protobuf.settings)
@@ -883,7 +888,7 @@ lazy val `persistence-javadsl` = (project in file("persistence/javadsl"))
     `persistence-testkit`,
     jackson,
     `cluster-javadsl`,
-    `cluster-extensions-javadsl`
+    `projection-javadsl`
   )
   .settings(runtimeLibCommon: _*)
   .settings(mimaSettings(since = version150): _*)
@@ -900,7 +905,7 @@ lazy val `persistence-scaladsl` = (project in file("persistence/scaladsl"))
     `persistence-testkit`,
     `play-json`,
     `cluster-scaladsl`,
-    `cluster-extensions-scaladsl`
+    `projection-scaladsl`
   )
   .settings(runtimeLibCommon: _*)
   .settings(mimaSettings(since = version150): _*)
@@ -1066,7 +1071,7 @@ lazy val `kafka-broker` = (project in file("service/core/kafka/server"))
     Dependencies.`kafka-broker`
   )
   .settings(runtimeLibCommon: _*)
-  .dependsOn(`api`, `persistence-core`, `kafka-client`)
+  .dependsOn(`api`, `persistence-core`, `projection-core`, `kafka-client`)
 
 lazy val `kafka-broker-javadsl` = (project in file("service/javadsl/kafka/server"))
   .enablePlugins(RuntimeLibPlugins)
@@ -1240,9 +1245,9 @@ lazy val `sbt-plugin` = (project in file("dev") / "sbt-plugin")
       val () = (publishLocal in `cluster-core`).value
       val () = (publishLocal in `cluster-javadsl`).value
       val () = (publishLocal in `cluster-scaladsl`).value
-      val () = (publishLocal in `cluster-extensions`).value
-      val () = (publishLocal in `cluster-extensions-scaladsl`).value
-      val () = (publishLocal in `cluster-extensions-javadsl`).value
+      val () = (publishLocal in `projection-core`).value
+      val () = (publishLocal in `projection-scaladsl`).value
+      val () = (publishLocal in `projection-javadsl`).value
       val () = (publishLocal in `immutables`).value
       val () = (publishLocal in `jackson`).value
       val () = (publishLocal in `logback`).value
