+++ cluster/core/src/main/resources/reference.conf
@@ -30,3 +30,11 @@ akka.actor.serialization-bindings {
   "akka.actor.Address"        = akka-misc
   "akka.remote.UniqueAddress" = akka-misc
 }
+
+
+# Cluster distribution settings
+lagom.persistence.cluster.distribution {
+
+  # Each entity is pinged at this interval. Each node will ping this often, so this interval can be quite long.
+  ensure-active-interval = 30s
+}
