--- cluster/extensions-javadsl/src/main/resources/reference.conf
@@ -1,1 +1,0 @@
-play.modules.enabled += com.lightbend.lagom.javadsl.cluster.projections.ProjectionRegistryModule
