--- cluster/extensions/src/main/resources/reference.conf
@@ -1,6 +1,0 @@
-# Cluster distribution settings
-lagom.persistence.cluster.distribution {
-
-  # Each entity is pinged at this interval. Each node will ping this often, so this interval can be quite long.
-  ensure-active-interval = 30s
-}
