+++ persistence/javadsl/src/main/scala/com/lightbend/lagom/internal/javadsl/persistence/ReadSideImpl.scala
@@ -14,7 +14,7 @@ import javax.inject.Singleton
 import akka.actor.ActorSystem
 import akka.cluster.Cluster
 import akka.stream.Materializer
-import com.lightbend.lagom.internal.cluster.projections.ProjectionRegistry
+import com.lightbend.lagom.internal.projection.ProjectionRegistry
 import com.lightbend.lagom.internal.persistence.ReadSideConfig
 import com.lightbend.lagom.internal.persistence.cluster.ClusterStartupTask
 import com.lightbend.lagom.javadsl.persistence._
