+++ persistence/scaladsl/src/main/scala/com/lightbend/lagom/scaladsl/persistence/PersistenceComponents.scala
@@ -6,12 +6,12 @@ package com.lightbend.lagom.scaladsl.persistence
 
 import akka.actor.ActorSystem
 import akka.stream.Materializer
-import com.lightbend.lagom.internal.cluster.projections.ProjectionRegistry
+import com.lightbend.lagom.internal.projection.ProjectionRegistry
 import com.lightbend.lagom.internal.persistence.ReadSideConfig
 import com.lightbend.lagom.internal.scaladsl.persistence.ReadSideImpl
 import com.lightbend.lagom.scaladsl.cluster.ClusterComponents
-import com.lightbend.lagom.scaladsl.cluster.projections.ProjectionComponents
-import com.lightbend.lagom.scaladsl.cluster.projections.Projections
+import com.lightbend.lagom.scaladsl.projection.ProjectionComponents
+import com.lightbend.lagom.scaladsl.projection.Projections
 import play.api.Configuration
 
 import scala.concurrent.ExecutionContext
