+++ projection/javadsl/src/main/resources/reference.conf
@@ -1,0 +1,1 @@
+play.modules.enabled += com.lightbend.lagom.javadsl.projection.ProjectionRegistryModule
